Added roomManager
