+++ app/frontend/src/chat/chat.viewModel.tsx
@@ -6,12 +6,8 @@ import { ChatContext } from "src/chat/chat.context";
 import { useRootViewModelContext } from "../root.context";
 import {
   DevError,
-  MessagePayload,
-  MessageType,
   UserListItem,
   ChatRoomPayload,
-  RoomType,
-  CreateRoomRequest,
   ChatRoomStatus
 } from "./chat.types";
 import {
@@ -22,6 +18,8 @@ import {
   handleNewMessageCreator,
   handleAddedToNewChatRoomCreator
 } from "./lib/socketHandler";
+import { handleSocketErrorResponse } from "./lib/helperFunctions";
+import { RoomManagerProvider, useRoomManager } from "./lib/roomManager";
 
 export interface ChatViewModelType extends ChatModelType {
   joinRoom: (roomName: string, password: string) => Promise<boolean>;
@@ -38,7 +36,7 @@ export interface ChatViewModelType extends ChatModelType {
 
 export const ChatViewModelProvider = ({ children }) => {
   /***********************/
-  /*   State Variables   */
+  /*   Model Variables   */
   /***********************/
   const chatModel = useChatModel();
   const {
@@ -47,41 +45,33 @@ export const ChatViewModelProvider = ({ children }) => {
     currentRoomName,
     setCurrentRoomName,
     setShowNewRoomSnackbar,
-    rooms,
-    setRooms,
-    setCurrentRoomMessages,
-    contextMenuData,
-    setContextMenuRoomsVisible
+    setCurrentRoomMessages
   } = chatModel;
 
   const { pageState, setPageState } = useRootViewModelContext();
 
   /**********************/
+  /*   Room Functions   */
+  /**********************/
+
+  const {
+    rooms,
+    setRooms,
+    updateRooms,
+    convertMessagePayloadToMessageType,
+    addMemberToRoom,
+    addChatRoom,
+    addMessageToRoom,
+    joinRoom,
+    sendRoomMessage,
+    createNewRoom,
+    leaveRoom,
+    changeRoomStatus
+  } = useRoomManager();
+
+  /**********************/
   /*   Util Functions   */
   /**********************/
-  const convertMessagePayloadToMessageType = (
-    messagePayload: MessagePayload
-  ): MessageType => {
-    const timestamp = new Date(messagePayload.timestamp);
-    const timestamp_readable = timestamp.toLocaleTimeString("en-US", {
-      hour: "numeric",
-      minute: "numeric",
-      hour12: true
-    });
-
-    return {
-      username: messagePayload.username,
-      roomId: messagePayload.roomName,
-      content: messagePayload.content,
-      timestamp_readable,
-      timestamp,
-      isOwn: messagePayload.username === tempUsername,
-      displayUser: true,
-      displayTimestamp: true,
-      displayDate: true,
-      avatar: rooms[messagePayload.roomName]?.avatars[messagePayload.username]
-    };
-  };
 
   // This function will be called when a room focus is changed.
   // If the previous room is the same as the current one, toggle the page state to PageState.Home
@@ -99,273 +89,6 @@ export const ChatViewModelProvider = ({ children }) => {
     setCurrentRoomName(roomName);
     setCurrentRoomMessages(rooms[roomName].messages);
     setPageState(PageState.Chat);
-  };
-
-  /**
-   *
-   * @param updateFn Callback function that takes in the previous rooms state and returns the new rooms state
-   */
-  const updateRooms = (updateFn) => {
-    setRooms((prevRooms) => {
-      const newRooms = { ...prevRooms };
-      updateFn(newRooms);
-      return newRooms;
-    });
-  };
-
-  // Helper function to handle errors
-  const handleSocketErrorResponse = (res: DevError | any): res is DevError => {
-    return (res as DevError).error !== undefined;
-  };
-
-  /**********************/
-  /*   Room Functions   */
-  /**********************/
-
-  // Add member to room
-  const addMemberToRoom = (roomName: string, member: UserListItem) => {
-    setRooms((prevRooms) => {
-      const newRooms = { ...prevRooms };
-      if (!newRooms[roomName]) {
-        console.log("In addMemberToRoom, room not found: ", roomName);
-        return newRooms;
-      }
-      newRooms[roomName].users[member.username] = member;
-      return newRooms;
-    });
-  };
-
-  const getChatRoomMembers = async (roomName: string) => {
-    return new Promise<{ [key: string]: UserListItem }>((resolve) => {
-      socket.emit(
-        "listUsers",
-        { chatRoomName: roomName },
-        (users: UserListItem[]) => {
-          const usersObj = users.reduce<{ [key: string]: UserListItem }>(
-            (acc, user) => {
-              acc[user.username] = user;
-              return acc;
-            },
-            {}
-          );
-          resolve(usersObj);
-        }
-      );
-    });
-  };
-
-  // Adds a new room to the rooms state variable
-  const addChatRoom = async (
-    chatRoomPayload: ChatRoomPayload
-  ): Promise<RoomType> => {
-    const userList = await getChatRoomMembers(chatRoomPayload.name);
-
-    // Validate the payload
-    if (!chatRoomPayload.name) {
-      console.log("In addChatRoom, invalid payload: ", chatRoomPayload);
-      return;
-    }
-
-    return new Promise<RoomType>((resolve) => {
-      const {
-        name,
-        status,
-        queryingUserRank,
-        latestMessage,
-        lastActivity,
-        avatars
-      } = chatRoomPayload;
-      const convertedLatestMessage = latestMessage
-        ? convertMessagePayloadToMessageType(latestMessage)
-        : undefined;
-
-      const newRoom = {
-        name: name,
-        status: status,
-        rank: queryingUserRank,
-        latestMessage: convertedLatestMessage,
-        messages: [],
-        lastActivity,
-        hasUnreadMessages: false,
-        avatars,
-        users: userList
-      };
-      updateRooms((newRooms) => {
-        if (!newRooms[name]) {
-          newRooms[name] = newRoom;
-        }
-        console.log("Added room to rooms: ", newRooms);
-        resolve(newRoom);
-      });
-    });
-  };
-
-  // FIXME: move to model?
-  const addMessageToRoom = (roomName: string, message: MessageType) => {
-    updateRooms((newRooms) => {
-      if (!newRooms[roomName]) {
-        console.log("addMessageToRoom: Room does not exist");
-      } else {
-        newRooms[roomName].messages.push(message);
-        newRooms[roomName].latestMessage = message;
-      }
-    });
-  };
-
-  const addMessagesToRoom = (roomName: string, messages: MessageType[]) => {
-    updateRooms((newRooms) => {
-      if (!newRooms[roomName]) {
-        console.log("addMessageSSSSSSSToRoom: Room does not exist");
-      } else {
-        newRooms[roomName].messages.push(...messages);
-        newRooms[roomName].latestMessage = messages[messages.length - 1];
-      }
-    });
-  };
-
-  // Create a new room
-  const createNewRoom = async (
-    roomName: string,
-    roomStatus: ChatRoomStatus,
-    roomPassword?: string
-  ): Promise<boolean> => {
-    const roomRequest: CreateRoomRequest = {
-      name: roomName,
-      status: roomStatus,
-      password: roomPassword,
-      owner: tempUsername
-    };
-    console.log("ChatPage: Creating new room", { ...roomRequest });
-
-    const response = await new Promise<DevError | ChatRoomPayload>((resolve) =>
-      socket.emit("createRoom", roomRequest, resolve)
-    );
-
-    if (handleSocketErrorResponse(response)) {
-      console.log("Error response from join room: ", response.error);
-      return false;
-    }
-
-    await addChatRoom(response as ChatRoomPayload);
-    return true;
-  };
-
-  // Join a room
-  const joinRoom = async (
-    roomName: string,
-    password: string
-  ): Promise<boolean> => {
-    const joinRoomPayload = { roomName, password, user: tempUsername };
-    const joinRoomRes = await new Promise<DevError | ChatRoomPayload>(
-      (resolve) => socket.emit("joinRoom", joinRoomPayload, resolve)
-    );
-
-    if (handleSocketErrorResponse(joinRoomRes)) {
-      console.log("Error response from join room: ", joinRoomRes.error);
-      alert(joinRoomRes.error);
-      return false;
-    }
-
-    const room = await addChatRoom(joinRoomRes as ChatRoomPayload);
-    selectRoom(room.name);
-
-    const messageRequest = { roomName, date: new Date(), pageSize: 50 };
-    const messagesRes = await new Promise<DevError | MessagePayload[]>(
-      (resolve) => socket.emit("getRoomMessagesPage", messageRequest, resolve)
-    );
-
-    if (handleSocketErrorResponse(messagesRes)) {
-      console.log("Error response from get room messages: ", messagesRes.error);
-      return false;
-    }
-
-    const messages = (messagesRes as MessagePayload[]).map((message) =>
-      convertMessagePayloadToMessageType(message)
-    );
-    addMessagesToRoom(roomName, messages);
-    setCurrentRoomName(roomName);
-
-    return true;
-  };
-
-  const leaveRoom = async (): Promise<boolean> => {
-    return new Promise<boolean>((resolve) => {
-      if (!contextMenuData) return;
-      const roomName = contextMenuData.name;
-      socket.emit(
-        "leaveRoom",
-        { roomName, username: tempUsername },
-        (response: DevError | string) => {
-          if (handleSocketErrorResponse(response)) {
-            console.log("Error response from leave room: ", response.error);
-            resolve(false);
-          }
-        }
-      );
-      setRooms((prevRooms) => {
-        const newRooms = { ...prevRooms };
-        delete newRooms[roomName];
-        return newRooms;
-      });
-      setContextMenuRoomsVisible(false);
-      if (currentRoomName === roomName) {
-        setCurrentRoomMessages([]);
-        setCurrentRoomName("");
-        setPageState(PageState.Home);
-      }
-      resolve(true);
-    });
-  };
-
-  const sendRoomMessage = async (
-    roomName: string,
-    message: string
-  ): Promise<boolean> => {
-    return new Promise<boolean>((resolve) => {
-      socket.emit(
-        "sendMessage",
-        {
-          sender: tempUsername,
-          roomName: roomName,
-          content: message
-        },
-        (res: DevError | string) => {
-          if (typeof res === "object" && res.error) {
-            console.log("Error response from send message: ", res.error);
-            resolve(false);
-          }
-        }
-      );
-      resolve(true);
-    });
-  };
-
-  const changeRoomStatus = async (
-    newStatus: ChatRoomStatus
-  ): Promise<boolean> => {
-    return new Promise<boolean>((resolve) => {
-      const roomName = contextMenuData.name;
-      console.log(`Changing room status of ${roomName} to ${newStatus}`);
-      // TODO: instead of sending only the status, send the whole room object
-      // TODO: if status is password, open a modal to ask for the password
-      const newRoom = rooms[roomName];
-      newRoom.status = newStatus;
-      socket.emit(
-        "updateChatRoom",
-        { newRoom },
-        (response: DevError | ChatRoomPayload) => {
-          if (handleSocketErrorResponse(response)) {
-            console.log("Error changing room status");
-            resolve(false);
-          }
-          console.log("Successfully changed room status");
-          updateRooms((newRooms) => {
-            newRooms[roomName] = newRoom;
-          });
-        }
-      );
-      resolve(true);
-    });
   };
 
   /**********************/
