+++ app/frontend/src/index.tsx
@@ -8,12 +8,13 @@ import { CssBaseline, ThemeProvider } from "@mui/material";
 import { ChatViewModelProvider } from "src/chat/chat.viewModel";
 import { ProfileViewModelProvider } from "./profile/profile.viewModel";
 import { RootViewModelProvider } from "./root.viewModel";
+import { SettingsViewModelProvider } from "./components/settings/settings.viewModel";
+import { RoomManagerProvider } from "./chat/lib/roomManager";
 
 import customTheme from "src/theme";
 import { RootView } from "./root.view";
 import { ErrorBoundary } from "react-error-boundary";
 import Fallback from "./components/error/error";
-import { SettingsViewModelProvider } from "./components/settings/settings.viewModel";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 const theme = customTheme();
@@ -24,14 +25,16 @@ root.render(
       <WebSocketProvider>
         <ThemeProvider theme={theme}>
           <RootViewModelProvider>
-            <ChatViewModelProvider>
-              <ProfileViewModelProvider>
-                <SettingsViewModelProvider>
-                  <CssBaseline />
-                  <RootView />
-                </SettingsViewModelProvider>
-              </ProfileViewModelProvider>
-            </ChatViewModelProvider>
+            <RoomManagerProvider>
+              <ChatViewModelProvider>
+                <ProfileViewModelProvider>
+                  <SettingsViewModelProvider>
+                    <CssBaseline />
+                    <RootView />
+                  </SettingsViewModelProvider>
+                </ProfileViewModelProvider>
+              </ChatViewModelProvider>
+            </RoomManagerProvider>
           </RootViewModelProvider>
         </ThemeProvider>
       </WebSocketProvider>
